+++ src/test/java/com/easypost/Fixtures.java
@@ -232,7 +232,7 @@ public abstract class Fixtures {
         If you need to re-record cassettes, increment the date below and ensure it is one day in the future,
         USPS only does "next-day" pickups including Saturday but not Sunday or Holidays.
          */
-        String pickupDate = "2023-06-08";
+        String pickupDate = "2023-12-08";
 
         fixture.put("min_datetime", pickupDate);
         fixture.put("max_datetime", pickupDate);
@@ -328,6 +328,6 @@ public abstract class Fixtures {
      * @return The default planned ship date
      */
     public static String plannedShipDate() {
-        return "2023-06-08";
+        return "2023-12-08";
     }
 }