+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Next Major Release
+
+- Removed `withCarbonOffset` parameter from `create`, `buy`, and `regenerateRates` functions of the Shipment service as EasyPost now offers Carbon Neutral shipments by default for free
+- Removes the undocumented `createAndBuy` function from the Batch service. The proper usage is to create a batch first and buy it separately
+
 ## v6.9.1 (2023-11-16)
 
 - Fixes a bug where address verification errors were not being properly deserialized